mengubah view login dan menambahkan fitur hak akses untuk setiap user

+++ app/Http/Controllers/DataBahanController.php
@@ -11,8 +11,20 @@ use RealRashid\SweetAlert\Facades\Alert;
 class DataBahanController extends Controller
 {
 
+    // public function __construct()
+    // {
+    //     // $this->authorizeResource(DataBahan::class, [
+    //     //     'viewAny',
+    //     //     'create',
+    //     //     'update',
+    //     //     'delete'
+    //     // ]);
+    // }
+
     public function index()
     {
+
+        $this->authorize(DataBahan::class, 'viewAny');
 
         // join tabel dengan tabel satuan
         $dataBahan = DataBahan::join('satuan', 'databahan.kd_satuan', '=', 'satuan.id_satuan')
@@ -26,6 +38,8 @@ class DataBahanController extends Controller
 
     public function create()
     {
+        $this->authorize('create', DataBahan::class);
+
         $kode = DataBahan::max('kd_bahan');
         $kode = (int) substr($kode, 3, 3);
         $kode = $kode + 1;
@@ -43,6 +57,8 @@ class DataBahanController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', DataBahan::class);
+
         $request->validate([
             'kd_bahan' => 'required|min:3|max:10',
             'nm_bahan' => 'required|min:3|max:50',
@@ -65,6 +81,8 @@ class DataBahanController extends Controller
     public function edit(DataBahan $dataBahan)
     {
 
+        $this->authorize('update', $dataBahan);
+
         $dataBahan = DB::table('databahan')->join('satuan', 'databahan.kd_satuan', '=', 'satuan.id_satuan')->select('databahan.*', 'satuan.nm_satuan')->where('kd_satuan', $dataBahan->kd_satuan)->first();
 
         $satuan = Satuan::all();
@@ -77,6 +95,7 @@ class DataBahanController extends Controller
 
     public function update(Request $request, DataBahan $dataBahan)
     {
+        $this->authorize('update', $dataBahan);
 
         // mengubah nama validasi
         $messages = [
@@ -110,6 +129,7 @@ class DataBahanController extends Controller
 
     public function destroy(DataBahan $dataBahan, Request $request)
     {
+        $this->authorize('delete', $dataBahan);
 
         $dataBahan->delete('kd_bahan', $request->kd_bahan);
         Alert::success('Data Bahan', 'Berhasil dihapus!');
